process_eval/requirements.txt
openai>=1.0.0
pyyaml>=6.0
tqdm>=4.60.0
python-dotenv>=1.0.0
